Added a main component to clear upp app.js
