+++ code/src/App.js
@@ -1,147 +1,6 @@
 import React, { useState } from "react";
-import { Header } from "./components/Header/Header";
-import { TravelLocation } from "components/TravelLocation/TravelLocation";
-import { AmountTravellers } from "components/AmountTravellers/AmountTravellers";
-import { PassengerInformation } from "components/PassengerInformation/PassengerInformation";
-import { ResponsibleForBooking } from "components/ResponsibleForBooking/ResponsibleForBooking";
-import { ShowBooking } from "components/ShowBooking/ShowBooking";
+import { Main } from "./components/Main/Main";
 
 export const App = () => {
-  const [step, setStep] = useState(4);
-  const [city, setCity] = useState("");
-  const [travelType, setTravelType] = useState("");
-  const [amountOfTravellers, setAmountOfTravellers] = useState(1);
-  const [nameList, setNameList] = useState([
-    {
-      firstName: "",
-      lastName: "",
-      passengerNumber: 1,
-    },
-    {
-      firstName: "",
-      lastName: "",
-      passengerNumber: 2,
-    },
-    {
-      firstName: "",
-      lastName: "",
-      passengerNumber: 3,
-    },
-    {
-      firstName: "",
-      lastName: "",
-      passengerNumber: 4,
-    },
-    {
-      firstName: "",
-      lastName: "",
-      passengerNumber: 5,
-    },
-  ]);
-  const [email, setEmail] = useState("");
-  const [telephoneNumber, setTelephoneNumber] = useState("");
-
-  const onSetFirstName = (firstName, passengerNumber) => {
-    let nameListCopy = [...nameList];
-    const passenger = nameListCopy[passengerNumber - 1];
-    passenger.firstName = firstName;
-    setNameList(nameListCopy);
-  };
-
-  const onSetLastName = (lastName, passengerNumber) => {
-    let nameListCopy = [...nameList];
-    const passenger = nameListCopy[passengerNumber - 1];
-    passenger.lastName = lastName;
-    setNameList(nameListCopy);
-  };
-
-  const onBtnClick = (e) => {
-    e.preventDefault();
-    setStep(step + 1);
-  };
-
-  const onSetCity = (e) => {
-    setCity(e.target.value);
-  };
-
-  const onSetTravelType = (travelType) => {
-    setTravelType(travelType);
-  };
-
-  const onSetAmountOfTravellers = (amountOfTravellers) => {
-    setAmountOfTravellers(amountOfTravellers);
-  };
-
-  const onSetEmail = (email) => {
-    setEmail(email);
-  };
-
-  const onSetTelephoneNumber = (telephoneNumber) => {
-    setTelephoneNumber(telephoneNumber);
-  };
-
-  console.log(city);
-  console.log(travelType);
-  console.log(amountOfTravellers);
-  console.log(nameList);
-  console.log(email);
-  console.log(telephoneNumber);
-  return (
-    <div className="content">
-      <Header step={step} />
-      <form className="form" onSubmit={onBtnClick}>
-        {step === 1 && (
-          <TravelLocation
-            onCityChanged={onSetCity}
-            onTravelTypeChanged={onSetTravelType}
-            city={city}
-            travelType={travelType}
-          />
-        )}
-
-        {step === 2 && (
-          <AmountTravellers
-            onAmountOfTravellers={onSetAmountOfTravellers}
-            amountOfTravellers={amountOfTravellers}
-          />
-        )}
-
-        {step === 3 && (
-          <PassengerInformation
-            amountOfTravellers={amountOfTravellers}
-            onChangeFirstName={onSetFirstName}
-            onChangeLastName={onSetLastName}
-          />
-        )}
-
-        {step === 4 && (
-          <ResponsibleForBooking
-            onChangeEmail={onSetEmail}
-            onChangeTelephoneNumber={onSetTelephoneNumber}
-          />
-        )}
-
-        {step === 5 && (
-          <ShowBooking
-            city={city}
-            nameList={nameList}
-            email={email}
-            telephoneNumber={telephoneNumber}
-            travelType={travelType}
-          />
-        )}
-
-        <button type="submit" className="btn submit" onSubmit={onBtnClick}>
-          Submit
-        </button>
-        <button
-          type="button"
-          className="btn previous-step"
-          onClick={() => setStep(step === 1 ? 1 : step - 1)}
-        >
-          Previous step
-        </button>
-      </form>
-    </div>
-  );
+  return <Main />;
 };
