+++ code/src/components/Main/main.css
@@ -1,0 +1,7 @@
+.submit {
+  margin-top: 20px;
+}
+
+.previous-step {
+  margin-top: 10px;
+}
