+++ code/src/index.css
@@ -33,11 +33,3 @@ h2 {
   margin: 30px 0 0 0;
   font-size: 20px;
 }
-
-.submit {
-  margin-top: 20px;
-}
-
-.previous-step {
-  margin-top: 10px;
-}
